+++ packages/website/components/customStorageForm/customStorageForm.js
@@ -22,86 +22,45 @@ const CustomStorageForm = ({ onClose }) => {
       <h3>Enterprise user inquiry</h3>
       <div className="fields">
         <div className="input-wrapper">
-          <label htmlFor="first-name">
-            First name<span className="text-red">*</span>
-          </label>
-          <input
-            type="text"
-            name="first-name"
-            id="first-name"
-            autoComplete="given-name"
-            // @ts-ignore
-            rules="required"
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
-          />
-        </div>
-
-        <div className="input-wrapper">
-          <label htmlFor="last-name">
-            Last name<span className="text-red">*</span>
-          </label>
-          <input
-            type="text"
-            name="last-name"
-            id="last-name"
-            // @ts-ignore
-            rules="required"
-            autoComplete="family-name"
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
-          />
-        </div>
-
-        <div className="input-wrapper">
-          <label htmlFor="email">
-            Email address associated with your web3.storage account
+          <label htmlFor="auth">
+            Please share the user authentication method (Github, Email) associated with your account.
             <span className="text-red">*</span>
           </label>
           <input
-            id="email"
-            name="email"
-            type="email"
-            autoComplete="email"
-            placeholder="name@email.com"
+            id="auth"
+            name="auth"
+            type="text"
+            placeholder="email address or github username"
             // @ts-ignore
-            rules="required|email"
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
+            rules="required"
           />
         </div>
         <div className="input-wrapper">
-          <label htmlFor="size">Estimated amount of storage in GiB</label>
-          <input
-            id="size"
-            name="size"
-            type="text"
-            placeholder="10,000"
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
-          />
+          <label htmlFor="size">
+            Please provide a ballpark estimate for your data volume (both in aggregate and over a given month).
+          </label>
+          <input id="size" name="size" type="text" placeholder="10Gib or ~1Gib/mo" />
         </div>
-        <div className="input-wrapper">
-          <label htmlFor="company_name">Company Name (if applicable)</label>
-          <input
-            id="company_name"
-            name="company_name"
-            type="text"
-            placeholder=""
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
-          />
+        <div className="input-wrapper full">
+          <label htmlFor="links">Please share links (Github, website, etc) for what you’re building.</label>
+          <textarea id="links" name="links" rows={2}></textarea>
         </div>
-        <div className="input-wrapper">
-          <label htmlFor="anything_else">Any other details we should know about</label>
-          <input
-            id="anything_else"
-            name="anything_else"
-            type="text"
-            placeholder=""
-            className="px-2 py-1 focus:ring-indigo-500 focus:border-blue block w-full border-black border-2 rounded-md"
-          />
+        <div className="input-wrapper full">
+          <label htmlFor="reading_data">
+            How do you plan on reading data uploaded to web3.storage (E.g., w3link gateway, other gateway, directly over
+            bitswap, etc.)? How frequently do you plan on reading data?
+          </label>
+          <textarea id="reading_data" name="reading_data"></textarea>
         </div>
-        <div className="sm:col-span-6">
-          <button type="submit" className="button outline-light cta Button">
-            Send
-          </button>
+        <div className="input-wrapper full">
+          <label htmlFor="anything_else">Is there any additional usage information we should know about?</label>
+          <textarea id="anything_else" name="anything_else"></textarea>
         </div>
+      </div>
+      <div className="sm:col-span-6">
+        <button type="submit" className="button outline-light cta Button">
+          Send
+        </button>
       </div>
     </form>
   );