+++ packages/website/components/messagebanner/messagebanner.js
@@ -95,7 +95,10 @@ export default function MessageBanner() {
   );
 
   return (
-    <section id="section_message-banner" style={{ height: bannerHeight }}>
+    <section
+      id="section_message-banner"
+      style={{ height: bannerHeight, display: !bannerContent.html.length ? 'none' : undefined }}
+    >
       <div ref={messageBannerRef} className={clsx('message-banner-wrapper', bannerHeight === '0px' ? 'mb-hidden' : '')}>
         <div className="grid-noGutter">
           <div className="col">