+++ packages/website/pages/callback/index.js
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { useQueryClient } from 'react-query';
 
-import Loading from 'components/loading/loading';
-import { redirectMagic, redirectSocial } from 'lib/magic.js';
+import { redirectMagic, redirectSocial } from '../../lib/magic.js';
+import Loading from '../../components/loading/loading.js';
 
 export function getStaticProps() {
   return {
@@ -16,13 +16,15 @@ export function getStaticProps() {
 const Callback = () => {
   const router = useRouter();
   const queryClient = useQueryClient();
-
+  const redirectUriQuery = router.query.redirect_uri;
+  const redirectUri =
+    (redirectUriQuery && Array.isArray(redirectUriQuery) ? redirectUriQuery[0] : redirectUriQuery) ?? '/account';
   useEffect(() => {
     const finishSocialLogin = async () => {
       try {
         await redirectSocial();
         await queryClient.invalidateQueries('magic-user');
-        router.push('/account');
+        router.push(redirectUri);
       } catch (err) {
         console.error(err);
         await queryClient.invalidateQueries('magic-user');
@@ -33,7 +35,8 @@ const Callback = () => {
       try {
         await redirectMagic();
         await queryClient.invalidateQueries('magic-user');
-        router.push('/account');
+
+        router.push(redirectUri);
       } catch (err) {
         console.error(err);
         await queryClient.invalidateQueries('magic-user');
@@ -46,7 +49,7 @@ const Callback = () => {
     if (router.query.provider) {
       finishSocialLogin();
     }
-  }, [router, router.query, queryClient]);
+  }, [router, router.query, queryClient, redirectUri]);
 
   // TODO handle errors
   return (