+++ packages/website/pages/login/index.js
@@ -14,12 +14,24 @@ const LoginType = {
   EMAIL: 'email',
 };
 
+/**
+ * Get the first thing
+ * @template T
+ * @param {T | T[] | undefined} thingOrThings
+ */
+function first(thingOrThings) {
+  if (Array.isArray(thingOrThings)) {
+    return thingOrThings[0];
+  }
+  return thingOrThings;
+}
+
 const Login = () => {
   // App query client
   const queryClient = useQueryClient();
 
   // App wide methods
-  const { push } = useRouter();
+  const { push, query } = useRouter();
 
   // Error states
   const [errors, setErrors] = useState(/** @type {{email?: string}} */ ({}));
@@ -36,10 +48,11 @@ const Login = () => {
   const onLoginWithEmail = useCallback(async () => {
     setErrors({ email: undefined });
     setIsLoggingIn(LoginType.EMAIL);
+    const finalRedirectUri = first(query.redirect_uri) ?? '/account';
     try {
-      await loginEmail(email || '');
+      await loginEmail(email || '', finalRedirectUri);
       await queryClient.invalidateQueries('magic-user');
-      push('/account');
+      push(finalRedirectUri);
     } catch (error) {
       setIsLoggingIn('');
 
@@ -48,7 +61,7 @@ const Login = () => {
       // @ts-ignore Catch clause variable type annotation must be 'any' or 'unknown' if specified.ts(1196)
       setErrors({ email: error.message });
     }
-  }, [email, push, queryClient]);
+  }, [email, push, queryClient, query.redirect_uri]);
 
   // Callback for github login logic
   const onGithubLogin = useCallback(async () => {